refactor: extract API types and experiment utilities out of api.py
Move EvaluationFrequency, TrainingStats, PreprocessedDataset, TrainingResults
to ludwig/api_types.py and _get_compute_description, get_experiment_description
to ludwig/experiment_utils.py. Re-export all symbols from api.py for backward
compatibility. Reduces api.py by ~180 lines.
